Using UrlHistoryList do display history list (on urientry edition).
[Issue]    https://bugs.tizen.org/jira/browse/TT-161
[Problem]  'Display URL from browser history that matches to given keyword.'
[Solution] UrlHistoryList is receiving list of matched entries and
creates a genlist. List is hidden, when list item is selected (with
cursor only for now) or focus changes from the uri entry. Created
genlist is not added to a focus chain yet (this will be delivered
soon).
[Verify] Typing words in uri entry should display list with matched
entries. Choosing item shoud load page in a new tab. Loading page or
focusing out from urientry should hide the list.

Signed-off-by: Adam Skobodzinski <a.skobodzins@partner.samsung.com>
Change-Id: Icb67a9b5c47556788d256613c832cdacb7620bf3